Moved queries from /components to /src
