+++ frontend/react/chatter-react/src/App.tsx
@@ -8,11 +8,7 @@ import React, { createContext, useEffect, useState } from "react";
 import { useCookies } from "react-cookie";
 import UserContext from "./components/Chatter/interface/UserContext";
 import Layout from "./components/Layout/Layout";
-import {
-  ADD_NEW_USER,
-  CREATE_LOBBY,
-  IS_LOBBY_EXISTING,
-} from "./components/Queries/App";
+import { ADD_NEW_USER, CREATE_LOBBY, IS_LOBBY_EXISTING } from "./queries/App";
 import LobbyModal from "./components/Lobby/LobbyModal";
 import IsLobbyExistingRequest from "./components/Chatter/interface/requests/IsLobbyExistingRequest";
 import IsLobbyExistingResponse from "./components/Chatter/interface/response/IsLobbyExistingResponse";
